+++ app/src/main/java/com/fibelatti/photowidget/preferences/UserPreferences.kt
@@ -15,5 +15,6 @@ data class UserPreferences(
     val defaultShape: String,
     val defaultCornerRadius: Float,
     val defaultOpacity: Float,
+    val defaultBlackAndWhite: Boolean,
     val defaultTapAction: PhotoWidgetTapAction,
 )