+++ app/src/main/java/com/fibelatti/photowidget/preferences/WidgetDefaultsViewModel.kt
@@ -39,6 +39,10 @@ class WidgetDefaultsViewModel @Inject constructor(
         userPreferencesStorage.defaultOpacity = value
     }
 
+    fun saveDefaultBlackAndWhite(value: Boolean) {
+        userPreferencesStorage.defaultBlackAndWhite = value
+    }
+
     fun saveDefaultTapAction(tapAction: PhotoWidgetTapAction) {
         userPreferencesStorage.defaultTapAction = tapAction
     }