+++ ui/src/main/java/com/fibelatti/ui/foundation/Modifier.kt
@@ -1,33 +1,11 @@
+@file:Suppress("unused")
+
 package com.fibelatti.ui.foundation
 
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.drawWithCache
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Rect
-import androidx.compose.ui.graphics.ColorFilter
-import androidx.compose.ui.graphics.ColorMatrix
-import androidx.compose.ui.graphics.Paint
-import androidx.compose.ui.graphics.drawscope.drawIntoCanvas
 
 inline fun Modifier.conditional(
     predicate: Boolean,
     ifTrue: Modifier.() -> Modifier,
     ifFalse: Modifier.() -> Modifier = { this },
 ): Modifier = if (predicate) then(ifTrue(Modifier)) else then(ifFalse(Modifier))
-
-fun Modifier.grayScale(): Modifier {
-    val saturationMatrix = ColorMatrix().apply { setToSaturation(0f) }
-    val saturationFilter = ColorFilter.colorMatrix(saturationMatrix)
-    val paint = Paint().apply { colorFilter = saturationFilter }
-
-    return drawWithCache {
-        val canvasBounds = Rect(Offset.Zero, size)
-        onDrawWithContent {
-            drawIntoCanvas {
-                it.saveLayer(canvasBounds, paint)
-                drawContent()
-                it.restore()
-            }
-        }
-    }
-}